+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Change default model to gpt-5.2
+- Upgrade Suntory to 0.12.0
 
 ## 1.0.0 - 2026-03-20
 ### Added